% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Industry_10.R
\docType{data}
\name{Industry_10}
\alias{Industry_10}
\title{Industry Portfolios}
\format{
A matrix (of size 252 x 10) containing daily returns of 10 industry portfolios.
}
\source{
\url{http://mba.tuck.dartmouth.edu/pages/faculty/ken.french/data_library.html}.
}
\usage{
data("Industry_10")
}
\description{
A matrix containing daily returns of 10 industry portfolios for the year 2014.
}
\note{
Data are available from Kenneth French's website.
}
\keyword{datasets}
